+++ app/src/main/java/com/example/cyberspace_info/autenticacao/adapter/INavegarTab.kt
@@ -1,4 +1,4 @@
-package com.example.cyberspace_info.login.view
+package com.example.cyberspace_info.autenticacao.adapter
 
 interface INavegarTab {
     fun mudarTab(posicaoAtual:Int)
