+++ app/src/main/java/com/example/cyberspace_info/autenticacao/registro/view/SignFragment.kt
@@ -1,4 +1,4 @@
-package com.example.cyberspace_info.login.view
+package com.example.cyberspace_info.autenticacao.registro.view
 
 import android.content.Context
 import android.os.Bundle
@@ -9,6 +9,8 @@ import android.view.ViewGroup
 import android.widget.Button
 import androidx.viewpager.widget.ViewPager
 import com.example.cyberspace_info.R
+import com.example.cyberspace_info.autenticacao.adapter.INavegarTab
+import com.example.cyberspace_info.autenticacao.view.LOGIN_FRAGMENT
 import com.google.android.material.tabs.TabLayout
 import com.google.android.material.textfield.TextInputEditText
