+++ app/src/main/java/com/example/cyberspace_info/autenticacao/view/AutenticacaoActivity.kt
@@ -1,9 +1,13 @@
-package com.example.cyberspace_info.login.view
+package com.example.cyberspace_info.autenticacao.view
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.viewpager.widget.ViewPager
 import com.example.cyberspace_info.R
+import com.example.cyberspace_info.autenticacao.adapter.INavegarTab
+import com.example.cyberspace_info.autenticacao.adapter.LoginAdapter
+import com.example.cyberspace_info.autenticacao.login.view.LoginFragment
+import com.example.cyberspace_info.autenticacao.registro.view.SignFragment
 import com.google.android.material.tabs.TabLayout
 
     const val LOGIN_FRAGMENT = 0
@@ -14,7 +18,7 @@ class LoginActivity : AppCompatActivity(), INavegarTab {
     lateinit var loginFragment: LoginFragment
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_login)
+        setContentView(R.layout.activity_autenticacao)
 
         val pager = findViewById<ViewPager>(R.id.viewPagerLogin)
         val tab = findViewById<TabLayout>(R.id.tabLayoutLogin)
