+++ app/src/main/res/layout/activity_autenticacao.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".login.view.LoginActivity"
+    tools:context=".autenticacao.view.LoginActivity"
     android:background="@color/colorBlueLightBackground">
