+++ app/src/main/res/layout/fragment_login.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical"
-    tools:context=".login.view.LoginFragment"
+    tools:context=".autenticacao.login.view.LoginFragment"
     android:padding="@dimen/dimen_16">
 
     <com.google.android.material.textfield.TextInputLayout
