+++ app/src/main/res/layout/fragment_sign.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".login.view.SignFragment"
+    tools:context=".autenticacao.registro.view.SignFragment"
     android:padding="@dimen/dimen_16">
 
     <com.google.android.material.textfield.TextInputLayout
